renamed setup option into pysetup
