+++ doc/option/pysetup/main.rst
@@ -1,5 +1,5 @@
-setup
-=====
+pysetup
+=======
 
 Add a 'setup.py' to your package to make it compliant with setuptools_. This
 will allow an easy distribution of your package. Since this option requires most
@@ -8,7 +8,7 @@ multiple 'manage add opt' commands.
 
 Quick setup::
 
-    (dvlpt)$ pmg add setup
+    (dvlpt)$ pmg add pysetup
     > intended versions [27]:
     ...
     (dvlpt)$ pmg regenerate
