+++ doc/readme.rst
@@ -53,7 +53,7 @@ already existing package:
  - :doc:`option/doc/main`: Add some documentation to your package
  - :doc:`option/test/main`: basic unitests using Nose_
  - :doc:`option/coverage/main`: add code coverage_ to the basic test configuration
- - :doc:`option/setup/main`: make your package distributable with setuptools (i.e. setup.py)
+ - :doc:`option/pysetup/main`: make your package distributable with setuptools (i.e. setup.py)
  - :doc:`option/data/main`: will guide through all the steps to add non python files to a package
  - :doc:`option/github/main`: will guide through all the step to safely store the package on Github_
  - :doc:`option/readthedocs/main`: step by step guide to have your documentation on ReadTheDocs_
@@ -84,7 +84,7 @@ To install a new option call the 'add' action::
 The script will perform different tasks sequentially:
 
  - Check if this option requires other options in order to be installed:
-   e.g. the 'setup' option requires all 'base, 'doc', 'test', 'license' and 'version'
+   e.g. the 'pysetup' option requires all 'base, 'doc', 'test', 'license' and 'version'
    in order to run properly.
  - Check if this option requires some extra packages in order to setup:
    e.g. the 'license' option depends on the lice_ package to function properly.
