+++ dvlpt_requirements.txt
@@ -1,4 +1,4 @@
-# {{pkglts setup.dvlpt_requirements,
+# {{pkglts pysetup.dvlpt_requirements,
 tox
 sphinx
 twine
