+++ src/pkglts_data/base/pysetup/requirements.txt
@@ -1,0 +1,1 @@
+# {{pkglts pysetup.install_requirements, }}
