+++ src/pkglts_data/base/pysetup/setup.cfg
@@ -1,7 +1,7 @@
 [global]
 # verbose=0
 
-# {{pkglts setup,
+# {{pkglts pysetup,
 [wheel]
 universal = 1
