--- src/pkglts_data/base/setup/dvlpt_requirements.txt
@@ -1,1 +1,0 @@
-# {{pkglts setup.dvlpt_requirements, }}
