--- src/pkglts_data/base/setup/requirements.txt
@@ -1,1 +1,0 @@
-# {{pkglts setup.install_requirements, }}
