+++ test/test_option/test_tox/test_handlers.py
@@ -2,6 +2,6 @@ from pkglts.option.tox.handlers import pyversions
 
 
 def test_handlers():
-    pkg_cfg = dict(setup={'intended_versions': ["27"]})
+    pkg_cfg = dict(pysetup={'intended_versions': ["27"]})
     txt = pyversions("txt", pkg_cfg)
     assert txt != "txt"
