+++ test/test_option/test_travis/test_handlers.py
@@ -6,7 +6,7 @@ def test_mapping():
 
 
 def test_handlers():
-    pkg_cfg = dict(setup={'intended_versions': ["27"]})
+    pkg_cfg = dict(pysetup={'intended_versions': ["27"]})
     txt = pyversions("txt", pkg_cfg)
     assert txt != "txt"
